+++ Controllers/EncoderController.h
@@ -58,7 +58,7 @@ enum {
 - (void)			stopEncoderTasksForDocument:(CompactDiscDocument *)document;
 
 - (BOOL)			hasTasks;
-- (unsigned)		countOfTasks;
+- (NSUInteger)		countOfTasks;
 
 // Action methods
 - (IBAction)		stopSelectedTasks:(id)sender;
